requirements.txt
future==0.18.2
python-telegram-bot==12.7